+++ src/lib/constants.js
@@ -1,4 +1,10 @@
 module.exports = {
+  // version to check when deciding if deck should be re-imported
+  // from scryfall. This should be changed whenever there's a new
+  // field added to the stored decks. For organizational puproses,
+  // the naming convention is YYYY-MM-DD__reason_for_change
+  VERSION: '2018-08-22__initial_version',
+
   DECK_LIST_TYPES: [
     'mainDeck',
     'commandZone',