+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # unreleased
 ## Fixes
 * Bump version of scryfall-client to v0.8.0
+* Include mechanism to refresh the cards from scryfall when changes are made to the API that require more information from the cards
 
 ## Features
 * Add automatic paging of search results